fix CL_OUT_OF_RESOURCES
Seems like things get a bit stricter when we use a TARGET buffer type.
So the buffer is created by grabbing the GetTargetId() as N.
Not sure what the difference would be, but it makes a difference.

app.cpp does null-check on screen, so it needs to be set to 0 for this to work and fully relinquish control to OpenCL for rendering the screen.

The program just hangs when I run it, but at least there is no OpenCL error. I am not sure if creating a global constant on the Game object is the best approach, and if I set screen = 0 at an ok point in the program.